working shell
